Fix 'Export Markdown' feature
