+++ src/components/Header/MenuButton.tsx
@@ -2,7 +2,6 @@ import MoreHorizIcon from '@mui/icons-material/MoreHoriz'
 import { Menu, MenuItem } from '@mui/material'
 import IconButton from '@mui/material/IconButton'
 import { useState } from 'react'
-import { toMarkdown } from '../../converterFunctions'
 import ThemeOption from './MenuOptions/ThemeOption'
 
 type Props = {
@@ -22,25 +21,6 @@ const MenuButton = ({ theme, toggleTheme, title, onUpload }: Props) => {
 
     const closeMenu = () => {
         setAnchor(null)
-    }
-
-    const onDownload = () => {
-        const fileName = `${title || 'NewFile'}.md`
-        const markdownText = toMarkdown(
-            document.getElementsByClassName('ql-editor')[0] as HTMLElement
-        )
-        downloadText(markdownText, fileName)
-    }
-
-    const downloadText = (text: string, fileName: string) => {
-        const element = document.createElement('a')
-        const file = new Blob([text], { type: 'text/plain;charset=utf-8' })
-        element.href = URL.createObjectURL(file)
-        element.download = fileName
-        element.hidden = true
-        document.body.appendChild(element)
-        element.click()
-        document.body.removeChild(element)
     }
 
     return (
