+++ src/components/Header/ToolbarContainer.jsx
@@ -1,9 +1,26 @@
 import FileOption from './ToolbarOptions/FileMenu/FileOption'
 
-const ToolbarContainer = ({ onUpload }) => {
+const ToolbarContainer = ({ onUpload, mdText, title }) => {
+    const onDownload = () => {
+        const fileName = `${title || 'NewFile'}.md`
+        //const markdownText = toMarkdown(mdText)
+        downloadText(mdText, fileName)
+    }
+
+    const downloadText = (text, fileName) => {
+        const element = document.createElement('a')
+        const file = new Blob([text], { type: 'text/plain;charset=utf-8' })
+        element.href = URL.createObjectURL(file)
+        element.download = fileName
+        element.hidden = true
+        document.body.appendChild(element)
+        element.click()
+        document.body.removeChild(element)
+    }
+
     return (
         <div>
-            <FileOption onUpload={onUpload} />
+            <FileOption onUpload={onUpload} onDownload={onDownload} />
         </div>
     )
 }
